ovs-poc:  include header file
In the original commit I forgot to include the new
header file.

However I will commit a follow up patch shortly to remove the hardcoded
values in the header file. They are not needed anymore because we can
query the hardware for the port number and also learn the mac address.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>